# backend/SanSo.Api.V4/ProgramV5Canonical.cs
using Npgsql;using SanSo.Api;using SanSo.Api.Modules;using SanSo.Api.V2;using SanSo.Api.V4;using SanSo.Import;
var builder=WebApplication.CreateBuilder(args);var cs=builder.Configuration.GetConnectionString("Postgres")??Environment.GetEnvironmentVariable("SANSO_POSTGRES");if(!builder.Environment.IsDevelopment()&&string.IsNullOrWhiteSpace(cs))throw new InvalidOperationException("Production requires SANSO_POSTGRES/ConnectionStrings:Postgres");if(cs is not null){builder.Services.AddSingleton(NpgsqlDataSource.Create(cs));builder.Services.AddSingleton<PostgresCommerceStore>();builder.Services.AddSingleton<PostgresLifecycleStore>();builder.Services.AddSingleton<PostgresImportStagingStore>();}builder.Services.AddSingleton<DemoStore>();builder.Services.AddSanSoModules();builder.Services.AddSingleton<IdentityService>();builder.Services.AddSingleton<AuditTrail>();builder.Services.AddSingleton<ReliableSync>();builder.Services.AddSingleton<FinancialLifecycle>();builder.Services.AddSingleton<TeamAndSupport>();builder.Services.AddSingleton<BillingLifecycle>();builder.Services.AddSingleton<SafeCopilot>();builder.Services.AddCors(x=>x.AddDefaultPolicy(p=>p.WithOrigins("http://localhost:5173","http://127.0.0.1:5173","http://127.0.0.1:5174","http://127.0.0.1:5176").AllowAnyHeader().AllowAnyMethod()));
var app=builder.Build();app.UseMiddleware<SafeProblemMiddleware>();app.UseCors();SanSo.Api.IdentityEndpoints.MapIdentity(app);app.MapSanSoModules();var importer=new CommerceFileImporter(new ImportRegistry());var memoryImport=new ImportConfirmationWorkflow();
string Tenant(HttpRequest r)=>r.HttpContext.Items["TenantId"]as string??r.Headers["X-Tenant-Id"].FirstOrDefault()??"tenant-an-nhien";
SessionPrincipal? Principal(HttpRequest r){if(r.HttpContext.Items["SessionPrincipal"]is SessionPrincipal principal)return principal;var tenant=r.Headers["X-Tenant-Id"].FirstOrDefault();var auth=r.Headers.Authorization.FirstOrDefault();if(tenant is null||auth?.StartsWith("Bearer ",StringComparison.OrdinalIgnoreCase)!=true)return null;try{return app.Services.GetRequiredService<IdentityService>().Authenticate(auth[7..],tenant);}catch(UnauthorizedAccessException){return null;}}
bool Db()=>cs is not null;PostgresCommerceStore Pg()=>app.Services.GetRequiredService<PostgresCommerceStore>();PostgresImportStagingStore PgImport()=>app.Services.GetRequiredService<PostgresImportStagingStore>();
app.MapGet("/health",async(CancellationToken ct)=>{if(!Db())return Results.Ok(new{status="degraded",database="not-configured",mode="development-demo",api="v5"});try{await using var c=await app.Services.GetRequiredService<NpgsqlDataSource>().OpenConnectionAsync(ct);await using var q=c.CreateCommand();q.CommandText="SELECT 1";await q.ExecuteScalarAsync(ct);return Results.Ok(new{status="ok",database="postgresql",api="v5"});}catch{return Results.Json(new{status="unhealthy",database="unavailable",api="v5"},statusCode:503);}});
app.MapPost("/api/imports/preview",async(HttpRequest request,CancellationToken ct)=>{if(!request.HasFormContentType)return Results.BadRequest(new{code="MULTIPART_REQUIRED"});var file=(await request.ReadFormAsync(ct)).Files.GetFile("file");if(file is null)return Results.BadRequest(new{code="FILE_REQUIRED"});if(file.Length>10*1024*1024)return Results.BadRequest(new{code="FILE_TOO_LARGE"});await using var stream=file.OpenReadStream();try{ImportPreview preview;var ext=Path.GetExtension(file.FileName).ToLowerInvariant();if(ext==".csv"&&file.ContentType is"text/csv"or"application/csv"or"application/vnd.ms-excel"or"application/octet-stream")preview=importer.PreviewCsv(Tenant(request),stream);else if(ext==".xlsx"&&file.ContentType is"application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"or"application/octet-stream")preview=importer.PreviewXlsx(Tenant(request),stream);else return Results.BadRequest(new{code="UNSUPPORTED_FILE_TYPE"});if(Db()){var stored=await PgImport().Stage(Tenant(request),preview,ct);return Results.Ok(new{stored.BatchId,stored.PreviewToken,stored.ExpiresAt,preview.Format,preview.TemplateVersion,preview.Checksum,preview.Delimiter,preview.Headers,preview.Rows,preview.Errors,preview.Duplicate,persisted=true});}var staged=memoryImport.Stage(Tenant(request),preview);return Results.Ok(new{batchId=(string?)null,staged.PreviewToken,staged.ExpiresAt,preview.Format,preview.TemplateVersion,preview.Checksum,preview.Delimiter,preview.Headers,preview.Rows,preview.Errors,preview.Duplicate,persisted=false});}catch(DecoderFallbackException){return Results.BadRequest(new{code="ENCODING_INVALID",hint="Save CSV as UTF-8"});}catch(InvalidDataException e){return Results.BadRequest(new{code=e.Message});}});
app.MapPost("/api/imports/confirm",async(HttpRequest request,ImportConfirmV5CanonicalRequest body,CancellationToken ct)=>{try{if(Db()){var actor=Principal(request)?.UserId??throw new UnauthorizedAccessException();var stored=await PgImport().Confirm(Tenant(request),body.PreviewToken,body.Checksum,actor,ct);return Results.Ok(new{stored.BatchId,stored.Checksum,stored.AcceptedRows,stored.RejectedRows,stored.Duplicate,persisted=true});}var memory=memoryImport.Confirm(Tenant(request),body.PreviewToken,body.Checksum);return Results.Ok(new{batchId=(string?)null,memory.Checksum,memory.AcceptedRows,memory.RejectedRows,memory.Duplicate,persisted=false});}catch(UnauthorizedAccessException){return Results.Unauthorized();}catch(InvalidOperationException e){return Results.BadRequest(new{code=e.Message});}});
app.MapPost("/api/imports/demo",(HttpRequest r,DemoStore demo)=>app.Environment.IsDevelopment()?Results.Ok(demo.ImportDemo(Tenant(r))):Results.NotFound());app.MapGet("/api/dashboard",async(HttpRequest r,DemoStore demo,CancellationToken ct)=>Db()?Results.Ok(await Pg().Dashboard(Tenant(r),ct)):Results.Ok(demo.Dashboard(Tenant(r))));app.MapGet("/api/orders",async(HttpRequest r,DemoStore demo,CancellationToken ct)=>Db()?Results.Ok(await Pg().Orders(Tenant(r),ct)):Results.Ok(demo.Orders(Tenant(r))));app.MapGet("/api/reconciliations/current",async(HttpRequest r,DemoStore demo,CancellationToken ct)=>Db()?Results.Ok(await Pg().CurrentReconciliation(Tenant(r),ct)):Results.Ok(demo.Reconcile(Tenant(r))));app.Run();public partial class Program{}record ImportConfirmV5CanonicalRequest(string PreviewToken,string Checksum);